requirements.txt
Django
djangorestframework
psycopg2
python-decouple
djangorestframework-simplejwt
